Move format command implementation out of app
